feat(stack): support env interpolation for x-okteto-identity-token expiration_seconds
The manifest-wide env expander keeps the YAML string tag on expanded scalars,
so expiration_seconds: ${EXP} previously failed to unmarshal into int64. Back
the field with a custom IdentityTokenExpiration type whose UnmarshalYAML accepts
either a YAML integer (the common explicit case) or a numeric string (an
env-expanded value), returning a clear error for non-numeric input. The
audience and mount_path fields already expand as strings.

Refs PROD-497

+++ pkg/cmd/stack/translate.go
@@ -493,7 +493,7 @@ func translateIdentityTokenVolume(svc *model.Service) *apiv1.Volume {
 	}
 	expiration := defaultIdentityTokenExpirationSeconds
 	if svc.IdentityToken.ExpirationSeconds != nil {
-		expiration = *svc.IdentityToken.ExpirationSeconds
+		expiration = int64(*svc.IdentityToken.ExpirationSeconds)
 	}
 	return &apiv1.Volume{
 		Name: identityTokenVolumeName,
